classification/requirements.txt
tqdm
pillow
scikit-learn
matplotlib
torch==1.7.0
torchvision==0.8.1
torchtoolbox
warmup_scheduler
